Проведена деструктуризация по модулям
