+++ js/origin-data.js
@@ -1,0 +1,51 @@
+const PHOTO_COUNT = 25;
+
+const LIKES_COUNT = {
+  MIN: 15,
+  MAX: 200
+};
+
+const DESCRIPTIONS = [
+  'Heart For The Future',
+  'Morning Obsession',
+  'Voodoo Moves',
+  'Time For My Good Times',
+  'A Moment Of A Man',
+  'Hurt By My Blues'
+];
+
+const AVATAR_COUNT = {
+  MIN: 1,
+  MAX: 6
+};
+
+const COMMENT_COUNT = {
+  MIN: 0,
+  MAX: 5
+};
+
+const COMMENT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const COMMENT_NAMES = [
+  'Iosif Duskin',
+  'Vikenti Rusky',
+  'Grigor Duboff',
+  'Stas Chernoff',
+  'Maks Lopatin',
+  'Larisa Romanoff',
+  'Kira Petroff',
+  'Ludmila Galkin',
+  'Akilina Ivanoff',
+  'Tatiana Falin',
+  'Evgenia Angeloff',
+  'Igor Minsky'
+];
+
+export { PHOTO_COUNT, LIKES_COUNT, DESCRIPTIONS, AVATAR_COUNT, COMMENT_COUNT, COMMENT_MESSAGES, COMMENT_NAMES };
